Move isValidSignature test to Wallet test
